Clean-up + better testing
