+++ internal/pkg/aws/aws_test.go
@@ -98,169 +98,3 @@ func TestGenerateSTSCredentials(t *testing.T) {
 		})
 	}
 }
-
-func Test_validateProfile(t *testing.T) {
-	type args struct {
-		file    []byte
-		profile string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			"Valid/NoProfileNameDefined",
-			args{
-				file: []byte(`
-				[default]
-				aws_access_key_id = blahblah
-				aws_secret_access_key = blahblah/blahblah`),
-				profile: "",
-			},
-			false,
-		},
-		{
-			"Valid/NonDefaultProfileNameDefined",
-			args{
-				file: []byte(`
-				[candycrush]
-				aws_access_key_id = blahblah
-				aws_secret_access_key = blahblah/blahblah`),
-				profile: "candycrush",
-			},
-			false,
-		},
-		{
-			"Invalid/InvalidCredentialsFile",
-			args{
-				file: []byte(`
-				-[default]
-				_aws_access_key_id = blahblah
-				a-ws_secret_access_key = blahblah/blahblah`),
-				profile: "",
-			},
-			true,
-		},
-		{
-			"Invalid/InvalidProfile",
-			args{
-				file:    []byte(""),
-				profile: "blah",
-			},
-			true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := validateProfile(tt.args.file, tt.args.profile); (err != nil) != tt.wantErr {
-				t.Errorf("validateProfile() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func Test_openFile(t *testing.T) {
-
-	type args struct {
-		path string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    []byte
-		wantErr bool
-	}{
-		{
-			"Invalid/NonExistentFile",
-			args{
-				path: "/some/unknown/path",
-			},
-			nil,
-			true,
-		},
-		{
-			"Valid/FileExists",
-			args{
-				path: "/knowntestfile.txt",
-			},
-			[]byte(`test`),
-			false,
-		},
-		{
-			"Valid/EmptyFileExists",
-			args{
-				path: "/emptyknowntestfile.txt",
-			},
-			[]byte(""),
-			false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := openFile(tt.args.path)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("openFile() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("openFile() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_validateToken(t *testing.T) {
-	type args struct {
-		token string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{
-			"Invalid/EmptyToken",
-			args{
-				token: "",
-			},
-			true,
-		},
-		{
-			"Invalid/ShortToken",
-			args{
-				token: "2321",
-			},
-			true,
-		},
-		{
-			"Invalid/NotNumbers",
-			args{
-				token: "23ss21",
-			},
-			true,
-		},
-		{
-			"Invalid/NotNumbersLong",
-			args{
-				token: "5364f'[73",
-			},
-			true,
-		},
-		{
-			"Valid/Token",
-			args{
-				token: "1234532",
-			},
-			false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := validateToken(tt.args.token); (err != nil) != tt.wantErr {
-				t.Errorf("validateToken() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
