+++ internal/pkg/aws/sts.go
@@ -2,11 +2,20 @@ package aws
 
 import (
 	"fmt"
+	"regexp"
 
 	"github.com/aws/aws-sdk-go/aws/awserr"
 
 	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/aws/aws-sdk-go/service/sts/stsiface"
+)
+
+const (
+	tokenValidationRegex string = "^[0-9]+$"
+)
+
+var (
+	tokenValidationRegexComplied = regexp.MustCompilePOSIX(tokenValidationRegex)
 )
 
 //STSIdentity represents the STS Identity
@@ -16,7 +25,23 @@ type STSIdentity struct {
 	UserID  string
 }
 
+func validateToken(token string) error {
+	if len(token) <= 5 {
+		return ErrInvalidToken
+	}
+	if !tokenValidationRegexComplied.MatchString(token) {
+		return ErrInvalidToken
+	}
+
+	return nil
+}
+
 func getSTSSessionToken(stsInstance stsiface.STSAPI, tokenCode string, mfaDeviceSerialNumber string) (*sts.Credentials, error) {
+
+	if err := validateToken(tokenCode); err != nil {
+		return nil, err
+	}
+
 	stsSession, err := stsInstance.GetSessionToken(&sts.GetSessionTokenInput{
 		TokenCode:    &tokenCode,
 		SerialNumber: &mfaDeviceSerialNumber,
